Update based on FYHenry suggestions
