+++ package.json
@@ -28,14 +28,10 @@
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@popperjs/core": "^2.11.8",
     "@types/aos": "^3.0.7",
-    "@types/jquery": "^3.5.30",
     "aos": "^2.3.4",
     "axios": "^1.6.8",
     "bootstrap": "^5.3.3",
     "boxicons": "^2.1.4",
-    "jquery": "^3.7.1",
-    "ngx-bootstrap": "^12.0.0",
-    "ngx-bootstrap-modal": "^2.0.1",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "webpack-dev-middleware": ">=6.1.2",
