+++ src/app/public-module/shared-public-module/authentication-by-mail.service.spec..ts
@@ -1,0 +1,20 @@
+import { Injectable } from '@angular/core';
+import { AuthenticationByMailSteps } from './authentication-by-mail.enum';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class EngineStepService {
+
+  currentStep:AuthenticationByMailSteps = AuthenticationByMailSteps.INIT;
+
+  constructor() { }
+  
+  getCurrentStep():AuthenticationByMailSteps{
+    return this.currentStep;
+  }
+
+  setCurrentStep(stepToSet:AuthenticationByMailSteps):void{
+    this.currentStep = stepToSet;
+  }
+}
