+++ src/app/public-module/shared-public-module/authentication-by-mail.service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { EngineStepService } from './engine-step.service';
+import { EngineStepService } from './authentication-by-mail.service.spec.';
 
 describe('EngineStepServiceService', () => {
   let service: EngineStepService;
