+++ src/app/public-module/shared-public-module/send-auth-code-and-act/send-auth-code-and-act.component.html
@@ -10,7 +10,7 @@
       <div *ngSwitchCase=1>
         <!-- Send authentication code -->
         <h3 class="mb-3">Demande d'authentification</h3>
-        <p style="font-size: smaller;">Nous devons nous assurez que vous êtes celui que vous prétendez.<br>
+        <p sclass="inote-small-comment">Nous devons nous assurez que vous êtes celui que vous prétendez.<br>
         En cliquant sur le bouton ci-dessous, vous allez recevoir sur votre adresse mail un code qui vous permettra de réinitialiser votre mot de passe
         </p>
         <button (click)="onAskingReceiveCode()" type="button" class="mt-2 btn btn-secondary border border-white">
@@ -26,7 +26,7 @@
         <!-- 3 - Input new password -->
         <div class="d-flex flex-column">
           <h3 class="mb-3">Saisie du nouveau mot de passe</h3>
-          <p style="font-size: smaller;">Nb : le code d'authentification vient d'être envoyé à l'adresse
+          <p class="inote-small-comment">Nb : le code d'authentification vient d'être envoyé à l'adresse
             <strong>{{this.emailWhereCodeSended}}</strong>
           </p>
           <div class="align-items-center input-group input-group-sm mb-3 mx-auto">
