--- src/app/public-module/shared-public-module/steps.enum.ts
@@ -1,8 +1,0 @@
-export enum Steps {
-    INIT=0,
-    REQUEST_AUTH_CODE= 1,
-    SEND_AUTH_CODE=2,
-    SEND_NEW_PASSWORD=3,
-    OPERATION_RESULT_STATUS=4,
-    FINISH=5,
-}
