+++ src/app/shared-module/footer/footer.component.html
@@ -4,56 +4,21 @@
             <div class="col-md-6 ">
                 <h3>Sections du site</h3>
                 <ul>
-                    <li><em class="bx bx-chevron-right"></em> <a href="#" class="text-decoration-none text-black-50 ">Contact</a></li>
-                    <li><em class="bx bx-chevron-right"></em> <a href="#" class="text-decoration-none text-black-50">Mentions légales</a></li>
-                    <li><em class="bx bx-chevron-right"></em> <a href="#" class="text-decoration-none text-black-50">Politique de confidentialité</a></li>
-                    <li><em class="bx bx-chevron-right"></em> <a href="#" class="text-decoration-none text-black-50">Cookies</a></li>
+                    <li><em class="bx bx-chevron-right"></em> <a class="text-decoration-none text-black-50 inote-link-normal">Contact</a></li>
+                    <li><em class="bx bx-chevron-right"></em> <a class="text-decoration-none text-black-50 inote-link-normal">Mentions légales</a></li>
+                    <li><em class="bx bx-chevron-right"></em> <a class="text-decoration-none text-black-50 inote-link-normal">Politique de confidentialité</a></li>
+                    <li><em class="bx bx-chevron-right"></em> <a class="text-decoration-none text-black-50 inote-link-normal">Cookies</a></li>
                 </ul>
             </div>
             <div class="col-lg-4 col-md-6 justify-content-around ">
                 <h4>Suivez-nous sur les réseaux sociaux</h4>
                 <div class="mt-3">
-                    <a href="#" class="linkedin"><em class="bx bxl-linkedin text-black-50"></em></a>
-                    <a href="#" class="twitter"><em class="bx bxl-twitter text-black-50"></em></a>
-                    <a href="#" class="instagram"><em class="bx bxl-instagram text-black-50"></em></a>
-                    <a href="#" class="facebook"><em class="bx bxl-facebook text-black-50"></em></a>
+                    <a class="linkedin"><em class="bx bxl-linkedin text-black-50 inote-link-normal"></em></a>
+                    <a class="twitter"><em class="bx bxl-twitter text-black-50 inote-link-normal"></em></a>
+                    <a class="instagram"><em class="bx bxl-instagram text-black-50 inote-link-normal"></em></a>
+                    <a class="facebook"><em class="bx bxl-facebook text-black-50 inote-link-normal"></em></a>
                   </div>
             </div>
         </div>
     </div>
 </footer>
-
-<!-- <div class=" d-flex flex row bg-body-tertiary">
-    <hr style="width: 50%;
-    border-style: solid;
-    border-color: white;
-    border-width: 1px;" class="ms-auto me-auto">
-    <div class="d-flex flex-row">
-        <div class="ms-auto col-4">
-            <div class="d-flex flex-column align-items-center">
-                <div class="d-flex flex-row text-black justify-content-center" style="gap:1rem">
-                    <a>Mentions légales</a>
-                    <a>Politique de confidentialité</a>
-                    <a>Cookies</a>
-                </div>
-                <div>© 2024 Inote, Inc</div>
-            </div>
-        </div>
-        <div class="col-4">
-            <div class="d-flex flex-column align-items-end me-2 mb-2">
-                <div>Suivez-nous sur les réseaux sociaux</div>
-                <div class="d-flex flex-row">
-                    <a href="https://www.facebook.com"><img src="../../../assets/facebook.svg" alt="logo facebook"
-                            style="width:2rem;"></a>
-                    <a href="https://www.instagram.com"><img src="../../../assets/instagram.svg" alt="logo instagram"
-                            style="width:2rem;"></a>
-                    <a href="https://www.linkedin.com"><img src="../../../assets/linkedin.svg" alt="logo linkedin"
-                            style="width:2rem;;"></a>
-                    <a href="https://www.twitter.com"><img src="../../../assets/x.svg" alt="logo x"
-                            style="width:2rem;;"></a>
-                   
-                </div>
-            </div>
-        </div>
-    </div>
-</div> -->
